refactor(transforms): rename proper-noun.transform to name.transform
The file now exports both sanitizeProperNoun and sanitizeName, making
"proper-noun" a misleading name. Renaming to name.transform better
reflects the shared purpose: sanitizing name fields across entity types
(users, groups, trips).
