+++ apps/api/src/modules/groups/dto/update-group.dto.ts
@@ -10,7 +10,7 @@ import {
 } from 'class-validator';
 
 import { GroupVisibility } from '@chamuco/shared-types';
-import { sanitizeName } from '@/common/transforms/proper-noun.transform';
+import { sanitizeName } from '@/common/transforms/name.transform';
 import { GroupCoverDto } from './group-cover.dto';
 
 export class UpdateGroupDto {
